fix loop with fluent api

+++ Shop/Shop.Models/DBModel/Interfaces/IReviewProduct.cs
@@ -5,7 +5,7 @@ namespace Shop.Models.Interfaces {
     public interface IReviewProduct {
         Product Product { get; set; }
         int ProductId { get; set; }
-        int ratingReview { get; set; }
+        double ratingReview { get; set; }
         string reviewDesc { get; set; }
         int ReviewProductId { get; set; }
         User User { get; set; }
