+++ Shop/Shop.Models/DTO/ReviewProductDTO.cs
@@ -10,7 +10,7 @@ namespace Shop.Models.DTO {
     public class ReviewProductDTO {
         public ProductDTO Product { set; get; }
         public int ProductId { set; get; }
-        public int ratingReview { set; get; }
+        public double ratingReview { set; get; }
         public string reviewDesc { set; get; }
         public int ReviewProductId { set; get; }
         public UserDTO User { set; get; }
